feat: enable AWS configmap mutation
