+++ pkg/provider/vault/configmap.go
@@ -36,12 +36,11 @@ func (m *mutator) MutateConfigMap(ctx context.Context, mutateRequest provider.Co
 	}
 	defer m.client.Close()
 
-	config := vaultinjector.Config{
+	injector := vaultinjector.NewSecretInjector(vaultinjector.Config{
 		TransitKeyID:     m.config.TransitKeyID,
 		TransitPath:      m.config.TransitPath,
 		TransitBatchSize: m.config.TransitBatchSize,
-	}
-	injector := vaultinjector.NewSecretInjector(config, m.client, nil, m.logger)
+	}, m.client, nil, m.logger)
 
 	mutateRequest.ConfigMap.Data, err = injector.GetDataFromVault(mutateRequest.ConfigMap.Data)
 	if err != nil {
@@ -50,11 +49,9 @@ func (m *mutator) MutateConfigMap(ctx context.Context, mutateRequest provider.Co
 
 	for key, value := range mutateRequest.ConfigMap.BinaryData {
 		if isValidPrefix(string(value)) {
-			binaryData := map[string]string{
+			mapData, err := injector.GetDataFromVault(map[string]string{
 				key: string(value),
-			}
-
-			mapData, err := injector.GetDataFromVault(binaryData)
+			})
 			if err != nil {
 				return err
 			}
